Finished refactoring VuFindIndexer to parts.
