--- import/index_java/src/org/vufind/index/VuFindIndexer.java
@@ -1,592 +1,0 @@
-package org.vufind.index;
-/**
- * Custom VuFind indexing routines.
- *
- * Copyright (C) Villanova University 2017.
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 2,
- * as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
- */
-
-import java.io.File;
-import java.io.FileReader;
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.io.OutputStreamWriter;
-import java.lang.StringBuilder;
-import java.text.ParseException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.LinkedHashSet;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-import java.util.Properties;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-import java.sql.*;
-import java.text.SimpleDateFormat;
-
-import org.apache.log4j.Logger;
-import org.marc4j.marc.ControlField;
-import org.marc4j.marc.DataField;
-import org.marc4j.marc.Record;
-import org.marc4j.marc.Subfield;
-import org.marc4j.marc.VariableField;
-import org.solrmarc.callnum.DeweyCallNumber;
-import org.solrmarc.callnum.LCCallNumber;
-import org.solrmarc.tools.CallNumUtils;
-import org.solrmarc.tools.SolrMarcIndexerException;
-import org.solrmarc.tools.Utils;
-import org.solrmarc.index.SolrIndexer;
-
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
-
-import org.w3c.dom.Document;
-import org.w3c.dom.Node;
-import org.w3c.dom.NodeList;
-
-/**
- *
- * @author Robert Haschart
- * @version $Id: VuFindIndexer.java 224 2008-11-05 19:33:21Z asnagy $
- *
- */
-public class VuFindIndexer extends SolrIndexer
-{
-    // Initialize logging category
-    static Logger logger = Logger.getLogger(VuFindIndexer.class.getName());
-
-    // Initialize VuFind database connection (null until explicitly activated)
-    private Connection vufindDatabase = null;
-    private UpdateDateTracker tracker = null;
-
-    // the SimpleDateFormat class is not Thread-safe the below line were changes to be not static 
-    // which given the rest of the design of SolrMarc will make them work correctly.
-    private SimpleDateFormat marc005date = new SimpleDateFormat("yyyyMMddHHmmss.S");
-    private SimpleDateFormat marc008date = new SimpleDateFormat("yyMMdd");
-
-    // Shutdown flag:
-    private boolean shuttingDown = false;
-
-    /**
-     * Default constructor
-     * @param propertiesMapFile the {@code x_index.properties} file mapping solr
-     *  field names to values in the marc records
-     * @param propertyDirs array of directories holding properties files
-     * @throws Exception if {@code SolrIndexer} constructor threw an exception.
-     */
-    public VuFindIndexer(final String propertiesMapFile, final String[] propertyDirs)
-            throws FileNotFoundException, IOException, ParseException {
-        super(propertiesMapFile, propertyDirs);
-    }
-
-    /**
-     * Log an error message and throw a fatal exception.
-     * @param msg message to log
-     */
-    private void dieWithError(String msg)
-    {
-        logger.error(msg);
-        throw new SolrMarcIndexerException(SolrMarcIndexerException.EXIT, msg);
-    }
-
-    /**
-     * Connect to the VuFind database if we do not already have a connection.
-     */
-    private void connectToDatabase()
-    {
-        // Already connected?  Do nothing further!
-        if (vufindDatabase != null) {
-            return;
-        }
-
-        String dsn = ConfigManager.instance().getConfigSetting("config.ini", "Database", "database");
-
-        try {
-            // Parse key settings from the PHP-style DSN:
-            String username = "";
-            String password = "";
-            String classname = "invalid";
-            String prefix = "invalid";
-            if (dsn.substring(0, 8).equals("mysql://")) {
-                classname = "com.mysql.jdbc.Driver";
-                prefix = "mysql";
-            } else if (dsn.substring(0, 8).equals("pgsql://")) {
-                classname = "org.postgresql.Driver";
-                prefix = "postgresql";
-            }
-
-            Class.forName(classname).newInstance();
-            String[] parts = dsn.split("://");
-            if (parts.length > 1) {
-                parts = parts[1].split("@");
-                if (parts.length > 1) {
-                    dsn = prefix + "://" + parts[1];
-                    parts = parts[0].split(":");
-                    username = parts[0];
-                    if (parts.length > 1) {
-                        password = parts[1];
-                    }
-                }
-            }
-
-            // Connect to the database:
-            vufindDatabase = DriverManager.getConnection("jdbc:" + dsn, username, password);
-        } catch (Throwable e) {
-            dieWithError("Unable to connect to VuFind database");
-        }
-
-        Runtime.getRuntime().addShutdownHook(new VuFindShutdownThread(this));
-    }
-
-    private void disconnectFromDatabase()
-    {
-        if (vufindDatabase != null) {
-            try {
-                vufindDatabase.close();
-            } catch (SQLException e) {
-                System.err.println("Unable to disconnect from VuFind database");
-                logger.error("Unable to disconnect from VuFind database");
-            }
-        }
-    }
-
-    public void shutdown()
-    {
-        disconnectFromDatabase();
-        shuttingDown = true;
-    }
-
-    class VuFindShutdownThread extends Thread
-    {
-        private VuFindIndexer indexer;
-
-        public VuFindShutdownThread(VuFindIndexer i)
-        {
-            indexer = i;
-        }
-
-        public void run()
-        {
-            indexer.shutdown();
-        }
-    }
-
-    /**
-     * Establish UpdateDateTracker object if not already available.
-     */
-    private void loadUpdateDateTracker() throws java.sql.SQLException
-    {
-        if (tracker == null) {
-            connectToDatabase();
-            tracker = new UpdateDateTracker(vufindDatabase);
-        }
-    }
-
-    /**
-     * Support method for getLatestTransaction.
-     * @return Date extracted from 005 (or very old date, if unavailable)
-     */
-    private java.util.Date normalize005Date(String input)
-    {
-        // Normalize "null" strings to a generic bad value:
-        if (input == null) {
-            input = "null";
-        }
-
-        // Try to parse the date; default to "millisecond 0" (very old date) if we can't
-        // parse the data successfully.
-        java.util.Date retVal;
-        try {
-            retVal = marc005date.parse(input);
-        } catch(java.text.ParseException e) {
-            retVal = new java.util.Date(0);
-        }
-        return retVal;
-    }
-
-    /**
-     * Support method for getLatestTransaction.
-     * @return Date extracted from 008 (or very old date, if unavailable)
-     */
-    private java.util.Date normalize008Date(String input)
-    {
-        // Normalize "null" strings to a generic bad value:
-        if (input == null || input.length() < 6) {
-            input = "null";
-        }
-
-        // Try to parse the date; default to "millisecond 0" (very old date) if we can't
-        // parse the data successfully.
-        java.util.Date retVal;
-        try {
-            retVal = marc008date.parse(input.substring(0, 6));
-        } catch(java.lang.StringIndexOutOfBoundsException e) {
-            retVal = new java.util.Date(0);
-        } catch(java.text.ParseException e) {
-            retVal = new java.util.Date(0);
-        }
-        return retVal;
-    }
-
-    /**
-     * Extract the latest transaction date from the MARC record.  This is useful
-     * for detecting when a record has changed since the last time it was indexed.
-     *
-     * @param record MARC record
-     * @return Latest transaction date.
-     */
-    public java.util.Date getLatestTransaction(Record record) {
-        // First try the 005 -- this is most likely to have a precise transaction date:
-        Set<String> dates = getFieldList(record, "005");
-        if (dates != null) {
-            Iterator<String> dateIter = dates.iterator();
-            if (dateIter.hasNext()) {
-                return normalize005Date(dateIter.next());
-            }
-        }
-
-        // No luck with 005?  Try 008 next -- less precise, but better than nothing:
-        dates = getFieldList(record, "008");
-        if (dates != null) {
-            Iterator<String> dateIter = dates.iterator();
-            if (dateIter.hasNext()) {
-                return normalize008Date(dateIter.next());
-            }
-        }
-
-        // If we got this far, we couldn't find a valid value; return an arbitrary date:
-        return new java.util.Date(0);
-    }
-
-    /**
-     * Get call numbers of a specific type.
-     * 
-     * <p>{@code fieldSpec} is of form {@literal 098abc:099ab}, does not accept subfield ranges.
-     *
-     *
-     * @param record  current MARC record
-     * @param fieldSpec  which MARC fields / subfields need to be analyzed
-     * @param callTypeSf  subfield containing call number type, single character only
-     * @param callType  literal call number code
-     * @param result  a collection to gather the call numbers
-     * @return collection of call numbers, same object as {@code result}
-     */
-    public static Collection<String> getCallNumberByTypeCollector(
-            Record record, String fieldSpec, String callTypeSf, String callType, Collection<String> result) {
-        for (String tag : fieldSpec.split(":")) {
-            // Check to ensure tag length is at least 3 characters
-            if (tag.length() < 3) {
-                //TODO: Should this go to a log? Better message for a bad tag in a field spec?
-                System.err.println("Invalid tag specified: " + tag);
-                continue;
-            }
-            String dfTag = tag.substring(0, 3);
-            String sfSpec = null;
-            if (tag.length() > 3) {
-                    sfSpec = tag.substring(3);
-            }
-
-            // do all fields for this tag
-            for (VariableField vf : record.getVariableFields(dfTag)) {
-                // Assume tag represents a DataField
-                DataField df = (DataField) vf;
-                boolean callTypeMatch = false;
-                
-                // Assume call type subfield could repeat
-                for (Subfield typeSf : df.getSubfields(callTypeSf)) {
-                    if (callTypeSf.indexOf(typeSf.getCode()) != -1 && typeSf.getData().equals(callType)) {
-                        callTypeMatch = true;
-                    }
-                }
-                System.err.println("callTypeMatch after loop: " + callTypeMatch);
-                if (callTypeMatch) {
-                    result.add(df.getSubfieldsAsString(sfSpec));
-                }
-            } // end loop over variable fields
-        } // end loop over fieldSpec
-        return result;
-    }
-    
-
-    /**
-     * Get call numbers of a specific type.
-     * 
-     * <p>{@code fieldSpec} is of form {@literal 098abc:099ab}, does not accept subfield ranges.
-     *
-     * @param record  current MARC record
-     * @param fieldSpec  which MARC fields / subfields need to be analyzed
-     * @param callTypeSf  subfield containing call number type, single character only
-     * @param callType  literal call number code
-     * @return set of call numbers
-     */
-    public static Set<String> getCallNumberByType(Record record, String fieldSpec, String callTypeSf, String callType) {
-        return (Set<String>) getCallNumberByTypeCollector(record, fieldSpec, callTypeSf, callType,
-                new LinkedHashSet<String>());
-    }
-
-    /**
-     * Get call numbers of a specific type.
-     * 
-     * <p>{@code fieldSpec} is of form {@literal 098abc:099ab}, does not accept subfield ranges.
-     *
-     * @param record  current MARC record
-     * @param fieldSpec  which MARC fields / subfields need to be analyzed
-     * @param callTypeSf  subfield containing call number type, single character only
-     * @param callType  literal call number code
-     * @return list of call numbers
-     */
-    public static List<String> getCallNumberByTypeAsList(Record record, String fieldSpec, String callTypeSf, String callType) {
-        return (List<String>) getCallNumberByTypeCollector(record, fieldSpec, callTypeSf, callType,
-                new ArrayList<String>());
-    }
-
-    /**
-     * Normalize LC numbers for sorting purposes (use only the first valid number!).
-     * Will return first call number found if none pass validation,
-     * or empty string if no call numbers.
-     *
-     * @param  record current MARC record
-     * @param  fieldSpec which MARC fields / subfields need to be analyzed
-     * @return sortable shelf key of the first valid LC number encountered, 
-     *         otherwise shelf key of the first call number found.
-     */
-    public String getLCSortable(Record record, String fieldSpec) {
-        // Loop through the specified MARC fields:
-        Set<String> input = getFieldList(record, fieldSpec);
-        String firstCall = "";
-        for (String current : input) {
-            // If this is a valid LC number, return the sortable shelf key:
-            LCCallNumber callNum = new LCCallNumber(current);
-            if (callNum.isValid()) {
-                return callNum.getShelfKey();   // RETURN first valid
-            }
-            if (firstCall.length() == 0) {
-                firstCall = current;
-            }
-        }
-
-        // If we made it this far, did not find a valid LC number, so use what we have:
-        return new LCCallNumber(firstCall).getShelfKey();
-    }
-
-    /**
-     * Get sort key for first LC call number, identified by call type.
-     * 
-     * <p>{@code fieldSpec} is of form {@literal 098abc:099ab}, does not accept subfield ranges.
-     *
-     *
-     * @param record  current MARC record
-     * @param fieldSpec  which MARC fields / subfields need to be analyzed
-     * @param callTypeSf  subfield containing call number type, single character only
-     * @param callType  literal call number code
-     * @return sort key for first identified LC call number
-     */
-    public String getLCSortableByType(
-            Record record, String fieldSpec, String callTypeSf, String callType) {
-        String sortKey = null;
-        for (String tag : fieldSpec.split(":")) {
-            // Check to ensure tag length is at least 3 characters
-            if (tag.length() < 3) {
-                //TODO: Should this go to a log? Better message for a bad tag in a field spec?
-                System.err.println("Invalid tag specified: " + tag);
-                continue;
-            }
-            String dfTag = tag.substring(0, 3);
-            String sfSpec = null;
-            if (tag.length() > 3) {
-                    sfSpec = tag.substring(3);
-            }
-
-            // do all fields for this tag
-            for (VariableField vf : record.getVariableFields(dfTag)) {
-                // Assume tag represents a DataField
-                DataField df = (DataField) vf;
-                boolean callTypeMatch = false;
-                
-                // Assume call type subfield could repeat
-                for (Subfield typeSf : df.getSubfields(callTypeSf)) {
-                    if (callTypeSf.indexOf(typeSf.getCode()) != -1 && typeSf.getData().equals(callType)) {
-                        callTypeMatch = true;
-                    }
-                }
-                // take the first call number coded as LC
-                if (callTypeMatch) {
-                    sortKey = new LCCallNumber(df.getSubfieldsAsString(sfSpec)).getShelfKey();
-                    break;
-                }
-            } // end loop over variable fields
-        } // end loop over fieldSpec
-        return sortKey;
-    }
-
-    /**
-     * The following several methods are designed to get latitude and longitude
-     * coordinates.
-     * Records can have multiple coordinates sets of points and/or rectangles.
-     * Points are represented by coordinate sets where N=S E=W.
-     *
-     * code adapted from xrosecky - Moravian Library
-     * https://github.com/moravianlibrary/VuFind-2.x/blob/master/import/index_scripts/geo.bsh
-     * and incorporates VuFind location.bsh functionality for GoogleMap display.
-     */
-
-
-    /**
-     * Update the index date in the database for the specified core/ID pair.  We
-     * maintain a database of "first/last indexed" times separately from Solr to
-     * allow the history of our indexing activity to be stored permanently in a
-     * fashion that can survive even a total Solr rebuild.
-     */
-    public UpdateDateTracker updateTracker(String core, String id, java.util.Date latestTransaction)
-    {
-        // Update the database (if necessary):
-        try {
-            // Initialize date tracker if not already initialized:
-            loadUpdateDateTracker();
-
-            tracker.index(core, id, latestTransaction);
-        } catch (java.sql.SQLException e) {
-            // If we're in the process of shutting down, an error is expected:
-            if (!shuttingDown) {
-                dieWithError("Unexpected database error");
-            }
-        }
-
-        // Send back the tracker object so the caller can use it (helpful for
-        // use in BeanShell scripts).
-        return tracker;
-    }
-
-    /**
-     * Get the "first indexed" date for the current record.  (This is the first
-     * time that SolrMarc ever encountered this particular record).
-     *
-     * @param record current MARC record
-     * @param fieldSpec fields / subfields to be analyzed
-     * @param core core name
-     * @return ID string
-     */
-    public String getFirstIndexed(Record record, String fieldSpec, String core) {
-        // Update the database, then send back the first indexed date:
-        updateTracker(core, getFirstFieldVal(record, fieldSpec), getLatestTransaction(record));
-        return tracker.getFirstIndexed();
-    }
-
-    /**
-     * Get the "first indexed" date for the current record.  (This is the first
-     * time that SolrMarc ever encountered this particular record).
-     *
-     * @param record current MARC record
-     * @param fieldSpec fields / subfields to be analyzed
-     * @return ID string
-     */
-    public String getFirstIndexed(Record record, String fieldSpec) {
-        return getFirstIndexed(record, fieldSpec, "biblio");
-    }
-
-    /**
-     * Get the "first indexed" date for the current record.  (This is the first
-     * time that SolrMarc ever encountered this particular record).
-     *
-     * @param record current MARC record
-     * @return ID string
-     */
-    public String getFirstIndexed(Record record) {
-        return getFirstIndexed(record, "001", "biblio");
-    }
-
-    /**
-     * Get the "last indexed" date for the current record.  (This is the last time
-     * the record changed from SolrMarc's perspective).
-     *
-     * @param record current MARC record
-     * @param fieldSpec fields / subfields to be analyzed
-     * @param core core name
-     * @return ID string
-     */
-    public String getLastIndexed(Record record, String fieldSpec, String core) {
-        // Update the database, then send back the last indexed date:
-        updateTracker(core, getFirstFieldVal(record, fieldSpec), getLatestTransaction(record));
-        return tracker.getLastIndexed();
-    }
-
-    /**
-     * Get the "last indexed" date for the current record.  (This is the last time
-     * the record changed from SolrMarc's perspective).
-     *
-     * @param record current MARC record
-     * @param fieldSpec fields / subfields to analyze
-     * @return ID string
-     */
-    public String getLastIndexed(Record record, String fieldSpec) {
-        return getLastIndexed(record, fieldSpec, "biblio");
-    }
-
-    /**
-     * Get the "last indexed" date for the current record.  (This is the last time
-     * the record changed from SolrMarc's perspective).
-     *
-     * @param record current MARC record
-     * @return ID string
-     */
-    public String getLastIndexed(Record record) {
-        return getLastIndexed(record, "001", "biblio");
-    }
-
-    /**
-     * Get access to the Logger object.
-     *
-     * @return Logger
-     */
-    public Logger getLogger()
-    {
-        return logger;
-    }
-
-    /**
-     * Normalize trailing punctuation. This mimics the functionality built into VuFind's
-     * textFacet field type, so that you can get equivalent values when indexing into
-     * a string field. (Useful for docValues support).
-     *
-     * Can return null
-     *
-     * @param record current MARC record
-     * @param fieldSpec which MARC fields / subfields need to be analyzed
-     * @return Set containing normalized values
-     */
-    public Set<String> normalizeTrailingPunctuation(Record record, String fieldSpec) {
-        // Initialize our return value:
-        Set<String> result = new LinkedHashSet<String>();
-
-        // Loop through the specified MARC fields:
-        Set<String> input = getFieldList(record, fieldSpec);
-        Pattern pattern = Pattern.compile("(?<!\b[A-Z])[.\\s]*$");
-        for (String current: input) {
-            result.add(pattern.matcher(current).replaceAll(""));
-        }
-
-        // If we found no matches, return null; otherwise, return our results:
-        return result.isEmpty() ? null : result;
-    }
-}
